requirements.txt
asgiref==3.4.1
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cymem==2.0.6
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.1
django-import-export==2.7.1
et-xmlfile==1.1.0
idna==3.3
Jinja2==3.0.3
langcodes==3.3.0
MarkupPy==1.14
MarkupSafe==2.0.1
murmurhash==1.0.6
numpy==1.22.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pathy==0.6.1
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.7
PyYAML==6.0
qrcode==7.3.1
requests==2.27.1
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlparse==0.4.2
srsly==2.4.2
tablib==3.1.0
thinc==8.0.13
tqdm==4.62.3
typer==0.4.0
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
wasabi==0.9.0
xlrd==2.0.1
xlwt==1.3.0
